requirements.txt
amqp==5.1.1
asgiref==3.6.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.2
django-autoslug==1.9.9
django-cors-headers==3.14.0
django-environ==0.10.0
django-polymorphic==3.1.0
django-rest-framework==0.1.0
django-rest-polymorphic==0.1.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
flake8==6.0.0
idna==3.4
kombu==5.2.4
mccabe==0.7.0
Pillow==9.5.0
prompt-toolkit==3.0.38
psycopg2==2.9.6
pycodestyle==2.10.0
pyflakes==3.0.1
PyJWT==2.6.0
pytz==2023.3
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
